requirements.txt
flask==2.0.1
python-dotenv==0.19.0
google-cloud-vision==2.4.2
openai==0.28.0
Werkzeug==2.0.1
Pillow==9.3.0
pytesseract==0.3.10 